% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download_timeseries.R
\name{download_timeseries}
\alias{download_timeseries}
\title{download timeseries data to local file destination}
\usage{
download_timeseries(site, variable, destination = tempdir(), ...)
}
\arguments{
\item{site}{a valid powstreams site (see \link{list_sites})}

\item{variable}{a valid variable name for timeseries data}

\item{destination}{string for a folder location}

\item{...}{additional arguments passed to \code{\link[mda.streams]{download_ts}},
for example \code{on_local_exists}}
}
\value{
file handle for downloaded file
}
\description{
download a timeseries file to a user-specified (or temp file) location
}
\examples{
download_timeseries(site = "nwis_50231500", variable = 'doobs_nwis')
}

