% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/explore_model.R
\name{explore_model}
\alias{explore_model}
\title{Explore powstreams metabolism models with graphs}
\usage{
explore_model(browse = TRUE)
}
\arguments{
\item{browse}{use browser for graph rendering}
}
\description{
Open interactive graphs for metabolism models
}

