% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/explore_timeseries.R
\name{explore_timeseries}
\alias{explore_timeseries}
\title{explore powstreams timeseries data on graphs}
\usage{
explore_timeseries(browse = TRUE)
}
\arguments{
\item{browse}{use browser for graph rendering}
}
\description{
open interactive graphs for timeseries data
}

