% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{export_data}
\alias{export_data}
\title{Export data to a local text file}
\usage{
export_data(data, file = "", keep.units = is.unitted(data),
  comment.char = "#", sep = "\\t", row.names = FALSE, quote = FALSE,
  ...)
}
\arguments{
\item{data}{a data.frame or unitted_data.frame of data}

\item{file}{the file to write to, as in \code{\link{write.table}}}

\item{keep.units}{logical. if data has units, should these be written as a
second line (just beneath the header) in the text file?}

\item{comment.char}{a single character, or "", with which to prepend the line
containing units information}

\item{sep}{the character string used to separate columns, as in 
\code{\link{write.table}}}

\item{row.names}{logical. whether to write row names to the file, as in 
\code{\link{write.table}}}

\item{quote}{logical. whether to place quotes around every data entry, as in 
\code{\link{write.table}}}

\item{...}{other arguments passed to \code{\link{write.table}}}
}
\description{
Export data to a local text file
}
\seealso{
\code{\link[unitted]{write_unitted}}, \code{\link{write.table}}
}

