% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_site_data}
\alias{get_site_data}
\title{get site data from a powstreams site}
\usage{
get_site_data(site_id, variable)
}
\arguments{
\item{site_id}{a unique identifier for a powstreams site}

\item{variable}{a variable of interest}
}
\value{
data from site according to variable match
}
\description{
returns data from site into R data.frame
}
\examples{
get_site_data('nwis_008192', variable = 'dissolved_oxygen')
}

