% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ts_metadata.R
\name{get_ts_metadata}
\alias{get_ts_metadata}
\title{Get information about timeseries variables and sources}
\usage{
get_ts_metadata()
}
\description{
Get information about timeseries variables and sources
}
