% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/google_site.R
\name{google_site}
\alias{google_site}
\title{Get a URL to the gage location on Google Maps}
\usage{
google_site(site, browse = TRUE)
}
\arguments{
\item{site}{a valid powstreams site (see \code{\link{list_sites}})}

\item{browse}{open up a browser tab to map location?}
}
\description{
Get a URL to the gage location on Google Maps
}
\examples{
\dontrun{
google_site("nwis_11126000")
google_site(c("nwis_01467200","nwis_09327000","nwis_351111089512501"), browse=FALSE)
}
}

