% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_timeseries.R
\name{list_timeseries}
\alias{list_timeseries}
\title{list timeseries data available for a particular site}
\usage{
list_timeseries(site, ...)
}
\arguments{
\item{site}{a valid powstreams site (see \link{list_sites})}

\item{...}{additional arguments passed to \code{\link[sbtools]{session_check_reauth}},
for example \code{username}}
}
\value{
a character vector of timeseries names, or NULL if none exists.
}
\description{
list timeseries data available for a particular site
}
\examples{
\dontrun{
list_timeseries(site = 'nwis_01018035')
}
}
\seealso{
\code{\link{download_timeseries}}
}

