% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_timeseries.R
\name{load_timeseries}
\alias{load_timeseries}
\title{load a timeseries into local R environment}
\usage{
load_timeseries(site, variable, session = NULL)
}
\arguments{
\item{site}{a valid powstreams site (see \link{list_sites})}

\item{variable}{a valid variable name for timeseries data}

\item{session}{a valid sciencebase session (see \code{\link[sbtools]{authenticate_sb}}).
Set \code{session = NULL} (default) for sites on sciencebase that are public.}
}
\value{
timeseries data.frame
}
\description{
load a timeseries into local R environment. Will download timeseries into
temp file, importa it into R as a data.frame, and delete the temporary file
}
\examples{
\dontrun{
dissolved_oxygen <- load_timeseries(site = 'nwis_01018035', variable = 'doobs')
}
}

