% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_folder.R
\name{locate_folder}
\alias{locate_folder}
\title{Find a high-level folder on ScienceBase}
\arguments{
\item{folder}{the folder to locate}

\item{format}{character indicating whether the folder should be returned as 
an ID or as a full URL}

\item{by}{character indicating how to search for the item: using tags ("tag",
the default and recommended option), by scanning the parent directory for 
the desired title ("dir"), or both in combination ("either")?}

\item{browser}{logical. Should the URL be opened in a browser?}
}
\description{
Find a high-level folder on ScienceBase
}
\examples{
\dontrun{
locate_folder("publications", format="url")
}
}

