% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ts_metadata.R
\name{search_ts_metadata}
\alias{search_ts_metadata}
\title{Search information about timeseries variables and sources}
\usage{
search_ts_metadata(var = NA, src = NA, var_src = NA, units = NA,
  metab_var = NA, var_descrip = NA, src_descrip = NA,
  ts_metadata = get_ts_metadata(), match_case = FALSE, fixed = TRUE)
}
\arguments{
\item{var}{text to search for in the var column}

\item{src}{text to search for in the src column}

\item{var_src}{text to search for in the var_src column}

\item{units}{text to search for in the units column}

\item{metab_var}{text to search for in the metab_var column}

\item{var_descrip}{text to search for in the var_descrip column}

\item{src_descrip}{text to search for in the src_descrip column}

\item{ts_metadata}{table of ts_metadata to search}

\item{match_case}{logical. should the case be matched exactly?}

\item{fixed}{logical. As in \code{\link{grepl}}, TRUE to require an exact 
match and FALSE to use regular expressions}
}
\description{
Search information about timeseries variables and sources
}
\examples{
search_ts_metadata(units='mgO2')
search_ts_metadata(var='GPP|ER',fixed=FALSE)
search_ts_metadata(src='estBest')
}

