% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/site_location.R
\name{site_location}
\alias{site_location}
\title{get site latitude and longitude}
\usage{
site_location(site)
}
\arguments{
\item{site}{a valid powstreams site (see \code{\link{list_sites}})}
}
\description{
get site latitude and longitude
}
\examples{
site_location("nwis_11126000")
site_location('nwis_08377200')
\dontrun{
site_location(list_sites(with_timeseries = c('doobs', 'wtr')))
}
}

