% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/upload_presentation.R
\name{upload_presentation}
\alias{upload_presentation}
\title{upload a presentation to sciencebase}
\usage{
upload_presentation(file, username)
}
\arguments{
\item{file}{a file path to the presentation for upload}

\item{username}{valid sciencebase username (will be prompted for password)}
}
\description{
upload a presentation to sciencebase
}

