% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/upload_publication.R
\name{upload_publication}
\alias{upload_publication}
\title{upload a publication to sciencebase}
\usage{
upload_publication(file, ...)
}
\arguments{
\item{file}{a file path to the presentation for upload}

\item{...}{additional parameters passed to \code{\link[sbtools]{session_check_reauth}}}

\item{username}{valid sciencebase username (will be prompted for password)}
}
\description{
upload a publication to sciencebase
}

