% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download_timeseries.R
\name{download_timeseries}
\alias{download_timeseries}
\title{download timeseries data to local file destination}
\usage{
download_timeseries(site, variable, destination = NULL, ...)
}
\arguments{
\item{site}{a valid powstreams site (see \link{list_sites})}

\item{variable}{a valid variable name for timeseries data}

\item{destination}{string for a folder location}

\item{...}{additional arguments passed to \code{\link[sbtools]{item_file_download}},
for example \code{overwrite_file} and \code{\link[sbtools]{session_check_reauth}},
for example \code{username}}
}
\value{
file handle for downloaded file
}
\description{
download a timeseries file to a user-specified (or temp file) location
}
\examples{
\dontrun{
download_timeseries(site = 'nwis_01018035', variable = 'doobs')
}
}

