% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/download_watershed.R
\name{download_watershed}
\alias{download_watershed}
\title{download watershed shapefile}
\usage{
download_watershed(site, destination = NULL, session = NULL, ...)
}
\arguments{
\item{site}{a valid powstreams site (see \link{list_sites})}

\item{destination}{string for a folder location}

\item{session}{a valid sciencebase session (see \code{\link[sbtools]{authenticate_sb}}).
Set \code{session = NULL} (default) for sites on sciencebase that are public.}

\item{...}{additional arguments passed to \code{\link[httr]{write_disk}} (e.g., \code{overwrite = TRUE})}
}
\value{
file handle for downloaded file
}
\description{
download a watershed shapefile to a user-specified location
}
\examples{
\dontrun{
download_watershed(site = 'nwis_01018035')
}
}

