% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/explore_map.R
\name{explore_map}
\alias{explore_map}
\title{explore powstreams site metadata on a map}
\usage{
explore_map(browse = TRUE)
}
\arguments{
\item{browse}{use browser for map rendering}
}
\description{
open an interactive map for site data
}

