% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_timeseries.R
\name{explore_timeseries}
\alias{explore_timeseries}
\title{Explore powstreams timeseries data on graphs}
\usage{
explore_timeseries(browse = TRUE)
}
\arguments{
\item{browse}{use browser for graph rendering}
}
\description{
Open interactive graphs for timeseries data
}
