% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{get_numeric_meta_columns}
\alias{get_numeric_meta_columns}
\title{Find which metadata columns are numeric.}
\usage{
get_numeric_meta_columns()
}
\description{
Helper function for shiny apps
}
\keyword{internal}

