% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_sites.R
\name{list_sites}
\alias{list_sites}
\title{get site data from a powstreams site}
\usage{
list_sites(with_timeseries = NULL, logic = c("all", "any"))
}
\arguments{
\item{with_timeseries}{NULL to return all sites, or character vector of
timeseries variables}

\item{logic}{how to join the constraints in with_timeseries, ...: is any of
the listed parameters sufficient, or do you need all of them to be
available for a site to qualify?}
}
\value{
a character vector of site IDs
}
\description{
returns data from site into R data.frame
}
\examples{
\dontrun{
list_sites() # get all sites

# get all sites with a given type of data
list_sites(with_timeseries = 'wtr_nwis')

# get all sites with both wtr and doobs data
list_sites(with_timeseries = c('wtr_nwis','doobs_nwis'))

# get all sites with EITHER wtr or doobs data (or both)
list_sites(with_timeseries = c('wtr_nwis','doobs_nwis'), logic='any')
}
}

