% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_sites.R
\name{list_sites}
\alias{list_sites}
\title{get site data from a powstreams site}
\usage{
list_sites(with_timeseries = NULL, ...)
}
\arguments{
\item{with_timeseries}{NULL for all sites, or a valid timeseries name}

\item{...}{Additional parameters supplied to \code{\link[sbtools]{session_check_reauth}}}
}
\value{
a character vector of powstreams sites according to user-specified parameters
}
\description{
returns data from site into R data.frame
}
\examples{
list_sites() # get all sites
\dontrun{
list_sites(with_timeseries = 'wtr') # get all sites with a given type of data
}
}

