% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list_sites}
\alias{list_sites}
\title{get site data from a powstreams site}
\usage{
list_sites(within_poly, session = NULL)
}
\arguments{
\item{within_poly}{FALSE for all sites, or a valid polygon}

\item{session}{a valid sciencebase session (see \code{\link{authenticate_sb}}).
Set \code{session = NULL} (default) for sites on sciencebase that are public.}
}
\value{
a character vector of powstreams sites according to user-specified parameters
}
\description{
returns data from site into R data.frame
}
\examples{
list_sites(within_poly = FALSE) # get all sites
\dontrun{
list_sites(within_poly = poly) # get all sites within a user specific polygon (inclusive)
}
}

