% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_timeseries.R
\name{list_timeseries}
\alias{list_timeseries}
\title{list timeseries data available for a particular site}
\usage{
list_timeseries(site, session = NULL)
}
\arguments{
\item{site}{a valid powstreams site (see \link{list_sites})}

\item{session}{a valid sciencebase session (see \code{\link[sbtools]{authenticate_sb}}).
Set \code{session = NULL} (default) for sites on sciencebase that are public.}
}
\value{
a character vector of timeseries names, or NULL if none exists.
}
\description{
list timeseries data available for a particular site
}
\examples{
list_timeseries(site = 'nwis_01018035')
}
\seealso{
\code{\link{download_timeseries}}
}

