% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_timeseries.R
\name{load_timeseries}
\alias{load_timeseries}
\title{load timeseries into local R environment}
\usage{
load_timeseries(site, variables, join.fun = inner_join, ...)
}
\arguments{
\item{site}{a valid powstreams site (see \code{\link{list_sites}})}

\item{variables}{a valid variable name for timeseries data, or a vector of
variable names}

\item{join.fun}{The \pkg{dplyr} function to use when successively adding the
second, third, ... variables in \code{variables}. Good options include
\code{\link{inner_join}}, \code{\link{full_join}}, and
\code{\link{left_join}}.}

\item{...}{additional parameters passed to \code{\link[sbtools]{session_check_reauth}}}
}
\value{
timeseries data.frame. Will be matched in time if multiple variables
  are used for \code{variable}
}
\description{
load timeseries into local R environment. Will download
  timeseries into temp file, import it into R as a data.frame, and delete the
  temporary file.
}
\examples{
\dontrun{
dissolved_oxygen <- load_timeseries(site = 'nwis_01018035', variable = 'doobs')
metab_input <- load_timeseries(site = 'nwis_01408500', variable = c('doobs','wtr'))
}
}
\seealso{
\code{\link{list_timeseries}}, \code{\link{list_sites}}
}

