% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/powstreams.R
\docType{package}
\name{powstreams}
\alias{powstreams}
\alias{powstreams-package}
\title{High-level functions for accessing the Continental Stream Metabolism data}
\description{
This package contains functions that make it easy for R users to use the
Powell Center project data in particular. For lower-level functions specific
to the Powell Center project, see \pkg{mda.streams}, and for still
lower-level functions generic to all ScienceBase tasks, see \pkg{sbtools}.
}
\section{Logging in}{


  Use the \code{\link{authenticate_sb}} command from \pkg{sbtools} to
  establish session credentials for working with ScienceBase.
}

\section{Navigating ScienceBase}{


  \itemize{

  \item \code{\link{list_sites}}

  \item \code{\link{list_tses}}

  }
}

\section{Reading data from ScienceBase}{


  \itemize{

  \item \code{\link{get_site_info}} - reports basic site information

  \item \code{\link{get_ts}} - loads data from SB into R

  \item \code{\link{download_ts}} - downloads data from SB into a file

  }
}

\section{interactive data exploration}{


  \itemize{

  \item \code{\link{explore_map}} - explore site metadata with an interactive
  map in a web browser

  \item \code{\link{explore_timeseries}} - explore timeseries data with
  interactive graphs in a web browser

  \item \code{\link{explore_model}} - explore a metabolism model with
  interactive graphs in a web browser

  }
}

\section{Other data types on ScienceBase}{


  \itemize{

  \item \code{\link{upload_presentation}} - uploads a presentation to the
  Presentations folder on SB

  \item \code{\link{upload_publication}} - uploads a presentation to the
  Publications folder on SB

  }
}

