% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/upload_presentation.R
\name{upload_presentation}
\alias{upload_presentation}
\title{upload a presentation to sciencebase}
\usage{
upload_presentation(file, ...)
}
\arguments{
\item{file}{a file path to the presentation for upload}

\item{...}{additional parameters passed to \code{\link[sbtools]{session_check_reauth}}}
}
\description{
upload a presentation to sciencebase
}

