% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_presentation.R
\name{upload_presentation}
\alias{upload_presentation}
\title{upload a presentation to sciencebase}
\usage{
upload_presentation(file)
}
\arguments{
\item{file}{a file path to the presentation for upload}
}
\description{
You must be logged into ScienceBase with \code{\link{login_sb}} to use this 
function
}
