% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_publication.R
\name{upload_publication}
\alias{upload_publication}
\title{Upload a publication to sciencebase}
\usage{
upload_publication(file)
}
\arguments{
\item{file}{a file path to the presentation for upload}
}
\description{
You must be logged into ScienceBase with \code{\link{login_sb}} to use this
function
}
